requirements.txt
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
Pillow>=10.0.0
fake-useragent>=1.4.0
urllib3>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
streamlit>=1.28.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0 